replace logic in customhook
