+++ src/widget/generalInformation/generalInfoForms/ui/GeneralInfoForms.module.scss
